[home] Move hero to a separate component
This is simply to simplify the main home component.
