+++ app/javascript/home/components/home_header.vue
@@ -68,7 +68,7 @@ export default {
 };
 </script>
 
-<style lang='scss'>
+<style lang='scss' scoped>
 @import "css/variables";
 
 #header {
